Update index.d.ts (#25119)
The done function takes an optional release (boolean) argument that tells whether the client should be released or put back in the pool. From: https://node-postgres.com/api/pool
`
The releaseCallback releases an acquired client back to the pool. If you pass a truthy value in the err position to the callback, instead of releasing the client to the pool, the pool will be instructed to disconnect and destroy this client, leaving a space within itself for a new client.
`