Remove the any from the typings for equalityComparer
Replaces the (a: any, b: any) signature for equalityComparer.  For observables the signature is `equalityComparer(a: T, b: T)`, but for computeds the more accurate signature is `equalityComparer(a: T | undefined, b: T)`, since the equality comparer is run with the initial computation of a computed, with `undefined` as the first argument.  For example:

```
const c = ko.pureComputed(() => 4);
c.equalityComparer = (a, b) => {
     console.log(`Comparing ${a} to ${b}`);
     return true;
};
c(); // Wake up the computed
// Prints "Comparing undefined to 4"
```