Cleanup 20190328 (#34297)
* Update project homepages

* Update project urls and fix protractor-browser-logs dependency

Its dependency on protractor was "latest" -- 6.0 -- which doesn't work because
protractor has a dependency on selenium-webdriver. In the Definitely
Typed repo, selenium-webdriver is treated as a
dependency on @types/selenium-webdriver because d.ts overrides .js.
But selenium-webdriver is at 3.0 while @types/selenium-webdriver is at
4.0, meaning that protractor's imports are broken for people (like DT)
who have @types/selenium-webdriver installed.

There are three ways to fix this:

1. Update @types/selenium-webdriver to 4.0. This is a big task.
2. Fix protractor-browser-log's protractor dependency to 5.4.2, right
before protractor updated to the latest version of selenium-webdriver.
This way the installed version of selenium-webdriver matches the version
of @types/selenium-webdriver in DT.
3. ???? Other tsconfig.json/package.jon wizardry? I don't think this is
a worthwhile path to explore.