"Ember: fix type for `Route#controllerFor`.
Previously, `Route#controllerFor` was defined as taking a name from the
controller registry, but it should accept the name of a route matching
either the current route or one of its parents in the route hierarchy.
We cannot (currently) constrain it to match the route hierarchy, so
simply using `string` is the best option we have and users can then
narrow it to a specific `Controller` subclass.