[stripe-v3] Update Stripe checkout types to include server implementation (#36856)
* Update Stripe checkout types to include server implementation

There are 2 ways to use Stripe Checkout: the client implementation and
the server implementation. The former requires more parameters, such as
redirect URLs and the products for the checkout session. In the server
implementation the session is created server-side so the only parameter
required for `redirectToCheckout` is the session ID.

* Move tslint rule disablement to specific line